post-PR fixes to linting and testing issues
